requirements.txt
anyio==4.2.0
certifi==2023.11.17
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pytest==7.4.4
python-dotenv==1.0.0
sniffio==1.3.0
tomli==2.0.1
typing_extensions==4.9.0
